[AUDIO] Enable submix in kitkat in XML HAL
BZ: 156679

Extended Audio policy was overriding usage of submix
and this is modified to take submix in/out device into
account.

From KK, Google has made submix as part of default devices
and once an active submix recording session is initated the
submix_out device is initiated in start input of base class.

Change-Id: I4edb32f7e245c99061aeb16eae9a6ce09a0cb626
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>