[Audio] VTSV:set LPAL criterion and implement policy changes
BZ: 167004

The Lpal device will be used as criterion by the routes.pfw.
So it has to be added in audio_criteria.conf.
This device is mapped on InputDevicesType.
The criterion AlwaysListening will be mapped on the parameter
vtsv_route (from AudioRoute) instead of vtsv_active (from
AudioPolicy).
Implement also specific policy changes for LPAL input source.

Change-Id: I4c51a0ba622eecb60c66738e85f5c1f49f183092
Signed-off-by: Philippe Afonso <philippex.afonso@intel.com>