[AUDIO] Fix bug in AEC when no echo reference is found
BZ: 166976

The HAL crashes when no echo reference is found for AEC algo. It
happens when trying to reset the echo reference the first time :
the root cause is that the _echoReference pointer is not set to
NULL by default, leading to a call to release_echo_reference()
even when it is not needed. Eventually a free() is called on a
null pointer.

This patch fixes the bug by setting _echoReference to NULL by
default, and by checking its value inside resetEchoReference().

Change-Id: Ic51da976296f0eb10e777c1f53c6f8dec01baeff
Signed-off-by: Frédéric Boisnard <fredericx.boisnard@intel.com>