Audio: Resolve a deadlock in audio effects management
BZ: 176996

When switching back and forth from Headset to IHF, sometimes
a deadlock situation occurs in the effects management.
This is because the lock used to protect operations on the
echo reference is the same as the one used to protect operations
on the stream.

This patch provides a dedicated lock for echo reference ressource.

Change-Id: If0811bf833d670e8ffb758e46f7122c95f5b84c4
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>