[AUDIO] Skype: input device is not set when forcing usage to speaker
BZ: 172061

Currently in a Skype call, switching from the UI from wsHS/BT to
IHF will not update the input device to built-in/back mic. It remains
to the microphone of the accessory. As the input source used for
Skype is AUDIO_SOURCE_MIC, the forced configuration to speaker and
forced usage FOR_COMMUNICATION is missing in the policy. Adding this
policy configuration to update the input device in case of forced usage
to speaker.

Set the input device to built-in mic, when the forced usage for
Skype (input source AUDIO_SOURCE_MIC) is set to IHF and the current
input device is wsHS or BT headset.

Change-Id: Ie4ccef38a8ea4a1ce3d821c5347b08bbd25ad654
Signed-off-by: Ben Romdhane, ImedX <imedx.ben.romdhane@intel.com>