[AUDIO] Dump audio Hw registers in case of PCM I/O write errors
BZ: 173565

When writing PCM frames, I/O errors can occur in driver or
audio firmware. Currently, no register information is printed.

In case of an I/O error is returned on write, dump the contents in
the shell of the audio hw registers containing relevant information for
debug purposes. This is achieved through a new parameter framework plugin
(debugfs) dedicated to the debug file system. This makes the debug
highly configurable.

Change-Id: I3ea9cd28135eeeed35f19f558819e1b5442dfce8
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>