[AudioComms] Reduce Lock range to necessary
BZ: 176558

There is an ANR that occurs during monkey test and MTBF:
Record thread calls setApplicabilityMask which autolocks in write mode
_streamlock. In this lock context, is called updateLatency, which calls
AudioRouteManager:getLatencyInUs who waits _routingLock to be unlocked.
_routingLock is currently locked in write mode by Playback thread in
reconsiderRouting. which calls AudioStream::isStarted who waits _streamLock
to be unlocked. So there is a deadlock.

To avoid that issue, we remove the Autolock in write mode of _streamlock in
setApplicabilityMask, and only lock it when writing. We add also a lock in read
mode in getApplicabilityMask.

Change-Id: I53382be838a9783080025d6b09950e10c52673cf
Signed-off-by: ptonnetx <pierre-louisx.tonnet@intel.com>