Manage read/write errors from the stream manager
BZ: 173565

Assert and error retry is done at tinyalsa lib level. This
is preventing from using debug information and librairies for
a single write or read error. Now in the stream manager, the
error can not be processed, as not returned by the tinyalsa lib.

Put back up the error management into the stream manager to be
able to process on reception of the first read/write error.
Tinyalsa stream library returns the error to the stream manager.
Also generating silence will be done for any error.

Change-Id: Iebf39a46b3e17a18bf9bb327e7beec7120f00478
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>