Replace include-path-for tinyalsa w/ hardcoded string
BZ: 187187

The "include-path-for" mechanism prevents hardcoding paths in the Makefiles.
However, it is based on a mapping, part of the build system and there is *no
mapping* defined for tinyalsa.  There is a mapping defined in Intel's parts of
the build system but, of course, this is only accessible within our branches.
This "feature" makes porting the Audio HAL outside of our mainline more
difficult.

Besides, tinyalsa's path is unlikely to change unless Google adds a mapping in
their base build system.  Hardcoding it has little drawback for now.

Change-Id: Id24c4d00b941720bd50d1851a4ac38136f171066
Signed-off-by: David Wagner <david.wagner@intel.com>