Correct libstlport usage in Android Makefiles
BZ: 187187

The proper way to use STLport as shared library is as simple as calling

    include external/stlport/libstlport.mk

at the end of the target's declaration.  It will update relevant LOCAL
variables.  If the way to use stlport changes, there's no need to change the
users.

This commit removes all previous usage and replaces it by this inclusion.

Change-Id: I261cb4496101348a97aea619d8a2c824150c774e
Signed-off-by: David Wagner <david.wagner@intel.com>