Remove the need for LOCAL_IMPORT_C_INCLUDE_DIRS_FROM_SHARED_LIBRARIES
BZ: 187187

This variable has been added by us in the build system but it isn't part of the
"official" build system.  It was introduced, as its name suggests, to get the
needed include directories for a given shared library without linking to it.

Instead, this patch removes this non-portable hack with another one:  beside
the shared library definition, a static library is defined without any source
file but exporting the list of include directories.  A library wishing to fetch
that list may simply link against that dummy static library.

Change-Id: Ic6cb9a6aed5f713a928356ed65c276f638137c22
Signed-off-by: David Wagner <david.wagner@intel.com>