[AUDIO_HALXML] KW issues in Stream Manager
BZ: 168594

KW sanity checks on HAL XML reveal 3 KW critical issues.

The first one is solved by this change.
The second one is not anymore present on mainline.
The third one is not applicable as its case is handled by assertion.

The KW issues are described in first comment of concerned BZ.

Change-Id: Ic71e164ffb6ae954da404d4da13120bc348d8681
Signed-off-by: Jorge Quintero <jorge.quintero@intel.com>