Audio: Check for audio device before handling streams
BZ: 189593

A tombstone is generated in case of routing error due to the
possible double deletion of a null stream, generating a SIGABRT.
This stream routing occurs even after an open error of the audio
device.

Do not attach or detach a stream if the audio device cannot be
opened.

Change-Id: I719fad45723681413a3c216ed8fb960bcb1786c8
Signed-off-by: Vincent Becker <vincentx.becker@intel.com>