Force a sync of criteria even if no need to reroute
BZ: 195236

Audio PFW criteria might not be updated immediately if no
reroute is required. It may lead not to apply some configuration
that are glitch safe and do not need to go through the 5 steps
routing.

This patch forces a criteria commit and apply configuration
even if no re routing is required.

Change-Id: I0bbfd82cb94d1cd9f3a8dc2cecaa4a770b39ebe3
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>