Enhance VoIP Band Type criterion management
BZ: 191183

The Band Type for VoIP is only set upon detection of the voice
communication input source. However, some VoIP application may
not make use of this input source. It prevents from configuring
correctly LPE pre processing according to the band.

This patch manages the CSV band in a more generic way, checking
the rate of the active input stream.

Change-Id: I62c604a6346f7d57f66dfd6a472e6d2018fededf
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>