Rename libutility to libpfw_utility
BZ: 207083

Opensourcing the parameter-framework core implied a lot of renaming.
In the open-source version, the libutility was renamed to libpfw_utility
a long time ago.
Since we merged the changes from the open-source version in our internal
tree, this plugin did not compile any more.

This patch fix the compile issue.

Change-Id: Id48b0125a775debb24d3cb7a47fa8cbeb75dffac
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>