Return an error instead of asserting if HAL init fails
If the audio HAL cannot locate the PFW files, it asserts.  This
causes the media server to crash and the boot to hang.  This patch
modifies the code to return an error instead of aborting.  This
allows the boot to complete.  The device will not however be able
to play any audio.

Issue: GMINL-1614
Change-Id: I32635a4cacac991535e24d309dad50a0cd46f12f
Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>
Reviewed-on: https://android.intel.com/216526
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Zajac, Guillaume <guillaume.zajac@intel.com>
Tested-by: Zajac, Guillaume <guillaume.zajac@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>
Reviewed-on: https://android.intel.com/280818
Tested-by: cactus <cactus@intel.com>
Reviewed-by: Lilja, Ola <ola.lilja@intel.com>