Avoid crash when AudioHAL cannot load MAMGR interface
BZ: 215151

When audio HAL fails to load the MAMGR interface, it crashes.
The platform state tries to read from all its value set the initial values.
As a result, it tries to read from modem proxy(ies) the state of the modem,
the status of the call and the band type.

This patch fixes this crash when failing to open the MAMGR interface by checking
the interface and returning default values when the MAMGR interface is not
available.

Change-Id: I3b15e579d2334205fe763f81199bf23e646a88b3
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>