Fix crash in route manager on startup.
BZ: 215151

When restarting the media server, some event reported to the
platform state may lead to reconsidering the routing before the
route manager service has been started.

This patch fixes this issue by avoiding to assert on such case and
just logging a warning. As a synchronisation is performed by AudioHAL
at the end of startup, no event will be lost or not taken into account.

Change-Id: I723d3ecb0196e17d7cbb8479482f3987ee7236d2
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>