Remove cast from bool to uint as convert from bool to string supported.
BZ: 213899

The conversion from boolean to string is now supported. It arbirarilly
return "true" or "false" strings.
This patch makes use of this convert specialization within ModemProxy to
format the setParameters. The consummer of the parameters supports the
conversion from "true" or "false" as also based on convert tool.
This patch also reports an error in case of failing conversion for mic
mute and update the mapping table to match the convention of convertTo.

Change-Id: Ie8916b99220bcbd0f9bc9422dba54b5a6596de37
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>