[WORKAROUND] Ignore Device None on output stream on AudioHAL.
When for example A2DP Headset is disconnected, the policy manager sends routing=0.
When the stream is started again, the policy does not force the routing as it
considers that the new device elected is the same than the previous one.
In fact, selecting device=AUDIO_DEVICE_NONE does not update the device member of output
descriptor, leading the policy to think that no changes happened on the routing.

This patch fixes the issue in the audio HAL by ignoring for output stream only the device
none parameter in order to remove the AOSP patch.
Revert this change once Google provides a clean fix in the policy.

Change-Id: I8f475f2497456487f92317ea02dd866e6c37624b
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>