Use standard library path for pfw plugin location
The route plugin used to be installed in
/system/lib/parameter-framework-plugins/Route.
Since the Parameter-framework is now able to load plugins from
standard library locations such as /system/lib(64)/,
there is no need for this specific location anymore.
This patch moves the install directory of libroute-subsystem.so to
the standard library path.

Issue: GMINL-1748

Change-Id: Ia83e0cc2a54cc4e8e95195465fde82b3cc6c3a33
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com/283357
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>
Tested-by: cactus <cactus@intel.com>
Tested-by: Dixon, CharlesX <charlesx.dixon@intel.com>
Reviewed-by: Lilja, Ola <ola.lilja@intel.com>