Update Modem Proxy to get more precise call status
AudioHAL used to be notified only when audio link with the modem
shall be activated by the MAMGR. The MAMGR used to simplify the call
status information into a boolean. The Noise Reduction should be activated
only when voice link is connected to reduce the convergence time.

This patch adds a new state in the call status, now it informs audio HAL
with three different states: AudioDetach, AudioActive and VoiceActive.

Change-Id: Icdcb72498593c667aca22e0fdb8e4b6570884bfa
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3860
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/283479