Update Modem Proxy to get more precise call status
AudioHAL used to be notified only when audio link with the modem
shall be activated by the MAMGR. The MAMGR used to simplify the call
status information into a boolean. The Noise Reduction should be activated
only when voice link is connected to reduce the convergence time.

This patch adds a new state in the call status, now it informs audio HAL
with three different states: AudioDetach, AudioActive and VoiceActive.

Depends-Change-Id: I3db2d3efd4b78dca3b9fe9ea6cd779e8862a2333
Change-Id: Icdcb72498593c667aca22e0fdb8e4b6570884bfa
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-4991
Signed-off-by: Frederic Boisnard <frederic.boisnard@intel.com>
Reviewed-on: https://android.intel.com/343033
Reviewed-by: Di Folco, Neil <neil.di.folco@intel.com>
Reviewed-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>
Tested-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>