Update audio hal configuration file to provide speech criterion
AudioHAL used to be notified only when audio link with the modem
shall be activated by the MAMGR. The MAMGR used to simplify the call
status information into a boolean. The Noise Reduction should be activated
only when voice link is connected to reduce the convergence time.

This patch updates the configuration file to translate the call state
parameter into route pfw "modem audio state" criterion and audio pfw
"speech audio state" criterion.

Change-Id: If69b3279f898453dc4f3377cb3aca8f5a8da5326
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-3860
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/283480