Safely manage input source conversion to mask
Since L-dessert, hotword input source is no more for the internal
use of the policy, it may also be sent to the audio HAL. AudioHAL
used to simpling convert the input source as a mask by shifting
with the input source value.
Hotword is outside the allowed shift to format a mask on 32 bits.

This patch allows to safely support hotword input source within Audio
HAL.

Issue: GMINL-4030

Change-Id: I6d5e51493350f47db9f9a7ae98b04f42f7124977
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com/300823
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Reviewed-by: De Chivre, Renaud <renaud.de.chivre@intel.com>