Fix gmock hard-coded include paths
The audio HAL functional tests are relying on the gmock framework, which is
not available within the AOSP.
For that, we have a copy of gmock in a private test repo.
The HAL test modules had hard coded include paths towards the gmock headers.
This is bad practice, since changing the tests repo's location will break the
build.

This patch removes the hard-coded includes since libgmock now exports
its headers.

Change-Id: I0dcc05bbeb8c3186d7bca5695933b1bfbf691d13
Depends-Change-Id: Iaa1d67d935a0d8551a991d5fcbbdf1911616f15a
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-16054
Signed-off-by: Mattijs Korpershoek <mattijsx.korpershoek@intel.com>
Reviewed-on: https://android.intel.com:443/316672