[Port from r4-dsda] Add getModemState to get modem tri-state values
For dual modem platforms, it is necessary to know the electrical
state of the modem, so that the shared audio bus connected to it
can be effectively used. This information is given to the audio
HAL by the Modem Audio Manager.

This new interface gets the modem tri-state Up, Down and Starting.
It will be used instead of isModemAlive that was returning a boolean
to tell if the modem was Up or Down.
Modem tri-states will be used to define the criteria ModemState
and Modem2State.

Change-Id: Ib4d91390d669181e65244d738cb8870276ceecd2
Orig-Change-Id: Ic84ab957d12d0741d1dd0260644ed78e64059b91
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-13126
Signed-off-by: Amine Kristou <aminex.kristou@intel.com>
Reviewed-on: https://android.intel.com:443/315622