Audio HAL 3.0: store the stream handle
In order to prepare the switch to new routing API, we need
to be able to identify the stream with their unique identifier
assigned by the policy.
This patch keeps track of the handle assigned by the policy
within streams objects and keep track of the opened streams
within primary Hw device.

Issue: GMINL-5451

Change-Id: If35fa51289673106d27a5dfca0c30b669c4bb209
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com/346042
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Bohlin, Johan <johan.bohlin@intel.com>