Manage collection of stream within Stream Mgr
In order to prepare the switch to new routing API, we need
to be able to identify the stream with their unique identifier
assigned by the policy.
This patch keeps track of the handle assigned by the policy
within streams objects and keep track of the opened streams
within primary Hw device.

Change-Id: I7dde070e424c95f81a57695e78bd2ce6e0f88032
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-21650
Reviewed-on: https://android.intel.com:443/331159