Fix assert due to mismatch number of bytes written returned to Flinger
On StreamOut::write() operation completion, Flinger checks for the
number of written bytes returned. The number of bytes written by the
Audio HAL to the StreamOut is either what was required by Flinger or
nothing. Trying to compute that number at the end of the write()
operation leads to an approximation error on the HAL side that induces
an assert in Flinger.

This patch fixes the potential assert by removing the computation of
the number of bytes written since StreamOut::write() operation always
write all bytes required or nothing at all.

This patch makes the write to return always the number of byte as it was
required (always success) as flinger will not manage the different value
correctly.

Change-Id: Ia651b0c05690617b9564afbed34296c6d115d6a7
Signed-off-by: Louis Le Gall <louis.le.gall@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-14274
Reviewed-on: https://android.intel.com:443/322306