[SIMPLIFIED CONF] Simplifiy criteria definition for PFW audio configuration files
Some criteria used by the PFW configuration can be simplified based on
the need of each platform.
In order to allow for minimal PFW configuration, these criteria can be
either removed from the structure and settings files, or moved to
relevant sections.

This patch simplifies the route_criteria.conf file to match the
changes made to the PFW config files as described above.

Change-Id: I81858e13baaff6128123e042875725c09aa9f229
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-6007
Signed-off-by: Apelete Seketeli <apeletex.seketeli@intel.com>
Signed-off-by: Louis Le Gall <louis.le.gall@intel.com>
Reviewed-on: https://android.intel.com/332205
Reviewed-on: https://android.intel.com/347699
Reviewed-by: Gaffie, Francois <francois.gaffie@intel.com>
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>