[SIMPLIFIED CONF] Allows to set criterion as numeric converted into string
Some criteria may be received as literal value, but they
may also be received as numerical value converted into string.
This patch allows to set the criterion as a string in both case.
It first checks if the literal matches a literal value of the
criterion type. If it fails, it try to convert it as a numerical
value. It also checks that this numerical value converted belongs
to the acceptable value of the criterion type.

Change-Id: Iaf5256377db3251c5a498ce99b9064667108a479
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-6007
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com/331802
Reviewed-on: https://android.intel.com/347694
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>