[SIMPLIFIED CONF] Allows to broadcast criterion to both PFW instances
Some criteria like output flags, input sources or pre proc effects
do not depends of android parameters and are knwon and set by AudioHAL.
They were intended to be used by Route PFW, and used to be forwarded
to audio PFW by the route pluging. This mecanism is heavy and expected
to be removed.

This patch allows to broadcast these criterion to both pfw instances,
depending on the behavior set by the HAL implemented in the
route_criteria.conf file.

Change-Id: I7ecbdcc0feebae61117b5479c8f8c7116a7ac5ba
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-6007
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com/331803
Reviewed-on: https://android.intel.com/347695
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>