Update framecount even if the stream is not routed
When the stream is not routed, the audio samples are trashed
in respect of the timeline by sleeping the amount of time that the HW
would have blocked the stream to consume these data. However, the
framecount is not updated, which is a mistake.

This patch updates the framecount whether the stream is routed or not

Change-Id: Ic2ffacc8d0e5188f6f178e4c91023533d8add701
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-18429
Reviewed-on: https://android.intel.com:443/335612