Allows to retrieve Mix ports from a given patch
This patch adds API to retrieve Mix Port from a given Audio Patch according
to the role it plays (i.e. source or sink). A port involves at most one mix
source and one mix sink. In order to route correctly, the stream manager
needs to extract from the patch the stream (so the mix port) involved in a patch
creation / deletion request.

Change-Id: I408ede4c4209c9eb525a363d33404725f722cfe5
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-7555
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>