Device Selection strategy within stream manager.
Platform state keeps track of streams.
It receives notification each time a stream is willing to start,
stop or change its device. it also receives notification on compress
stream on the same events even if managed in a different HAL.
This patch enforces the device selection amoung stream according to
their activity.
If the primary stream is active (whatever in parallel of deep, low
latency or compress), its devices are selected.
If the primary is not active and no other stream are active, its devices
are selected.
If primary is not active and another stream(s) is (are) the devices of
these streams.

Change-Id: I861d5b29e2ed4efffba59e0373d422d4c30168b3
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-7555
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>