Correct AudioPatch memory leak within AudioHAL
The AudioPatch collection is increasing, AudioHAL has a memory leak.
It is due to the fact that releaseAudioPatch was never called.

Whatever the application of new parameter succeed or not, once
a patch has been created or released successfully, the audio HAL must returns
OK to audio flinger in order to let flinger respectively add or remove the patch
to / from its internal patch collection.

This patch fixes this memory leak issue.

Change-Id: I670848b5de39e0b9bb0534237a24ba260015e59e
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-8745
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com/360964
Reviewed-by: Gonzalve, Sebastien <sebastien.gonzalve@intel.com>
Tested-by: Esclafer, CatherineX <catherinex.esclafer@intel.com>