Enforce Selection of output device according to mute stream attribute
When a primary stream is played on multiple devices (e.g. for alarm), any
other stream is muted. When the alarm is finished, the other streams are unmuted.
However, only when the 3sec inactivity time is expiring, the audio policy will
restore the output device to the original single one. During this time, the other
stream may play on dual devices.

This patch is enforcing the selection of output device within primary audio hal
to take this policy behavior into account and avoid playing an output stream
temporarily on multiple devices.

Change-Id: I5400ff6a4e458d834e995072ae63ff93f6c3b7ee
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-14461
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/409563