Improve patch management within Audio HAL
Exploring the way the audio policy is managing audio patches leads
to improve the management of patches within AudioHAL.
A patch may connects a source device to a since device but also a sink
mix as well. We need to take this into account in order to be ready
to route voice call the "non-legacy" way.

Change-Id: I62dec895728ca2ccbf3782d39668acb824946035
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-14624
Signed-off-by: Francois Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/395855