Gives routing priority for Direct stream
This patch gives to Direct Stream a higher priority upon routing
process prior non direct stream in case that 2 streams requested the same
devices.
Audio policy does not handle this kind of conflict. For example, both main
stream and multi channel stream may be running at the same time and requesting
to output on HDMI. This conflict is managed by giving a higher priority to the
direct stream, which gives the best user experience.

Change-Id: I51fa8caf014a1b7e517e1f4f23d86b97a01722cd
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7451
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/427893