allows to retrieve supported capabilities from a stream
This patch allows to retrieve the supported capabilities
from a stream (rate, format and channel mask).
It is intented to be used by the policy in case of a profile
declares capabilities as dynamic.
In this case, the policy may request from the stream its supported
rates, format and / or channel mask.

Change-Id: Iae2f73da150bf0c09bf029f564c807f227c6863a
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7451
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/428384