manage devices per stream in order to prepare route pfw removal
This patch restores the devices allocation per stream and gives
the way to route manager to associate a stream and a route according
to the stream devices and not according to the whole set of active
devices.
It prepares the removal of the route pfw rules for route applicability
which is a more and more a predictive problem to solve.

Change-Id: I85f3e7a146ceb99c5916c95bfc7d6d9852ab228b
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7451
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/432135