[WORKAROUND OF AOSP PATCH] Force discover of capabilities on first request
Ugly WA: the policy tries to open a stream to retrieve the capabilities BEFORE
broadcasting capabilities CONNECT message used by us to trig the discovery of the
capabilities. So, as capabilities are still empty, try to populate it...
Remove it once policy patch broadcasting device state before loading the profile
has been merged.

Change-Id: If6e1b4612572b4ab279a819c31f164ee9b3fbb64
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-7451
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/432693