Wrong Selection of Output Device
Currently, when primary and fast stream are both active in parallel
the output device of the fast stream is always selected.
ie: touch sound (Fast) and Phone ringtone (Primary), the selected
output device is Headphones and it has to be Headphones and Speaker.

This patch adds the check of Fast output flag to select appropriate
output devices.

Change-Id: Id44658770df88a361eec86ed4bcfc9ca8d7cd9e4
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-5694
Signed-off-by: Chokri ZAGROUBA <chokrix.zagrouba@intel.com>
Reviewed-on: https://android.intel.com:443/438363