audio stream manager: support pause / resume for direct streams.
When switching from one song to another, AudioFlinger may pause the stream
instead of stopping it. It prevents from using the HDMI card, which is a bottleneck
between stereo and multichannel direct streams. In order to allow playing the
next song that may be a non multichannel direct stream, we need to free the HW
ressource.

Change-Id: I09a7bc0fec51abf419450e6f3716896c915c3a79
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-12448
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/450516