[REVERT ME] Consider only last patch received for output device selection
While using routing APIs 3.0, we faced an issue on device selection.
In case of wHS connected, while music is played on wHS, if an alarm
is played, music is muted, alarm is broadcasted on speaker and wHS.
When the alarm is finished, music is unmuted, but the stream vehiculating
the alarm is still active until the inactivity timer expires.
From patch point of view, the audio HAL has 2 active patches:
- primary output mix to speaker + wHS
- music mix to wHS
As a result, we will make the wrong decision, routing music to speaker
to wHS.

This patch workarounds this issue by considering the output device from
the latest patch updated by the policy in case of source mix port only.

Change-Id: Ie5e752926688faf74b10fc3ab0dfae1de7898695
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-19662
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/465286