Fix audio stream destruction procedure
When deactivate BT while WsHS is connected, streams are not rerouted to WsHS
and we have the MediaServer tombstone "Pure virtual function called".
This crash occurs when we are deleting a stream that has not yet been stopped
by AudioFlinger.

Calling function "setStandby" in StreamOut.cpp/StreamIn.cpp classes instead
of parent class "Stream.cpp" fixes the issue.

Change-Id: Ia612b829bb1c19eda7a31b2a8da569ee60caf9b6
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-32346
Signed-off-by: Siala Mohamed AymenX <mohamed.aymenx.siala@intel.com>
Reviewed-on: https://android.intel.com:443/508885