Route manager: fix routing of concurrent streams.
Some concurrent streams may share the same attribute (use case, flags,
devices...).
One can be started and the other one in any order.
If the second stream is started then the first one is stopped, the parameters
of the platform state are identical and the route manager does not go into
the reconsider routing sequence.
As the stream is stopped, Android FW can legitimately delete the stream.
The next rerouting will lead to unroute a stream that was deleted.

This patch reconsider the routing inconditionnaly.

Change-Id: I64f04a0d430794ffd6a65ac4cd342125d5e0c628
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-35060
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/521208