Implement recover mechanism through uevent
This patch allows audio hal to listen to crash and recovery uevent
sent by the DSP.
On crash uevent, the audio subsystem is declared as not available
and all streams will be detached (playback will trash data, record will
record silence).
On recovery uevent, the streams are re connected if still active.

Change-Id: Ibe30be4c6ce874fdc46074aa5ed6d87a7bc60f26
Tracked-On: https://jira01.devtools.intel.com/browse/OAM-35077
Signed-off-by: François Gaffie <francois.gaffie@intel.com>
Reviewed-on: https://android.intel.com:443/521359