audio_hal: Parse the route attribute from criterion element
Usually the audio backend routes are enabled/disabled by
the setParameters function. Previously the Route PFW converted
the parameters to the applicable status of the backend route.
For example:
	supDomain: AuxStream

		supDomain: Capture

			confType: Yes
				AuxInSelected Is True

			domain: Applicable
				conf: Yes
					/Route/route_mgr/routes/aux_stream/
                    route_capture/status/applicable = 1

				conf: No
					/Route/route_mgr/routes/aux_stream/
                    route_capture/status/applicable = 0

Then the applicable status is converted to the disabling or
enabling of backend route in the AudioRouteManger and Audio
PFW. After removing the Route PFW, this ability is lost.

This patch adds a route attribute in the Parameter class,
and it means the parameter affects the route. When the parameter
is set by the AudioManager::setParameters, the AudioRouteManager
enables or disables the route affected by the parameters.

This patch also parses the route attribute from criterion element
in audio_criteria.xml and saves it into Parameter class.

The audio_criteria.xml defines the mapping relationship
between android paramters and intel parameters of audio PFW.
The criterion element is as following:

    <criterion name="AuxInSelected" type="BooleanType"
        default="False"
        parameter="aux_in_selected"
        mapping="false:False,true:True"
        route="auxin_to_speaker"/>

The 'auxin_to_speaker' is a route name from
audio_policy_configuration.xml. It is a backend route.

Change-Id: I4885a9ab9981c52501bc95e42e23fee661c096bb
Tracked-On:
Signed-off-by: yuanwu <yuan.wu@intel.com>
Reviewed-by: Mengdong Lin <mengdong.lin@linux.intel.com>
Reviewed-by: Libin Yang <libin.yang@intel.com>