audio_hal: Replace DevicePort struct with the new defined DevicePort class
An AudioRoute includes the audio source ports and sink
ports. AudioRouteManager needs the device ports and mix
ports information when parsing the AudioRoute from the
audio_policy_configuration.xml. In order to save the
different types of ports, make the changes as following:

a) Replace the DevicePort struct with DevicePort class
defined in AudioPort.hpp;
b) Replace the DevicePorts struct with AudioPorts class
defined in AudioPort.hpp;

Change-Id: I6ca4f5f08f1c6f33c4f751d062f66bf4e49af99c
Tracked-On:
Signed-off-by: yuanwu <yuan.wu@intel.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Mengdong Lin <mengdong.lin@linux.intel.com>
Reviewed-by: Libin Yang <libin.yang@intel.com>