audio_hal: Parse the name attribute of route element.
The patch is for parsing the route name from the
audio_policy_configuration.xml.
Before removing the Route PFW, the route name was
defined and generated by the Route PFW. For example:

<ComponentType Name="PlatformRoutes">
    <Component Name="media" Type="StreamRoutePlayback"
    Mapping="Card:@ SOUND_CARD_NAME@,Amend1:Media,
    Device:0,Ports:"/>
    <Component Name="aux_stream" Type="RouteCapture"
    Mapping="Amend1:AuxStream,Ports:"/>
</ComponentType>

The 'Media' and 'AuxStream' after 'Ammend1:' were the
route names. Actually the AuxStream is the backend route,
The Audio PFW used the route name in the audio pfw files
as following.

- The opened playback routes collection 'OpenedPlaybackRoutes'
  may include 'Media'.
- The opened capture routres collection 'OpenedCaptureRoutes'
  may include 'AuxStream'.

After removing the Route PFW, the mixPort element of
audio_policy_configuration.xml is parsed as AudioStreamRoute.
The name attribute of a mixPort element is used as the route
name. The name sharing is good enough for stream routes that
always have a mix port. However, backend routes have not mix
ports but device ports since they don't go through the
AudioFlinger. So the backend routes don't have a name for
identification. In addition, The StreamRouteCollection of
AudioRouteManager also doesn't save the audio backend routes.
Consequently, the AudioRouteManger cannot control any backend
routes.

Solution:
In order to support the AudioBackendRoute, add the
name attribute in the route element of
audio_policy_configuration.xml, so user can explicitly
define the name for a route, no matter it's a stream route
or backend route.
For example:
    <route type="mix" name="Media"
        sink="Speaker,Headset" source="primary_output"/>

The patch will parse the config file and store "Media' as
the route name.

Change-Id: Ida4ea05631c69c64de53f864a90be2aeb6958339
Tracked-On:
Signed-off-by: yuanwu <yuan.wu@intel.com>
Reviewed-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Reviewed-by: Mengdong Lin <mengdong.lin@linux.intel.com>
Reviewed-by: Libin Yang <libin.yang@intel.com>