audio_hal: Define the AudioBackendRoute class
Currently the AudioRouteManager can parse and control
the stream routes. But we lack support for backend
routes. Backend routes are used to describe hostless
audio paths likes AUX-IN to Speaker, or BT module to
speaker/DMIC etc. An audio backend route only includes
the devicePorts as sources and sinks. This patch
introduces an AudioBackendRoute class as the derived
class of AudioRoute.

Change-Id: I1179e49ae452e7be2b42a7784417152b560c6dc9
Tracked-On:
Signed-off-by: yuanwu <yuan.wu@intel.com>
Reviewed-by: Mengdong Lin <mengdong.lin@linux.intel.com>
Reviewed-by: Libin Yang <libin.yang@intel.com>