audio_hal: Add BACKEND_ROUTE in RouteType
Currently only stream route type is supported. This
patch adds the BACKEND_ROUTE type in the enum RouteType.

In order to support the audio backend route,
we also make following changes:
1) Add 'OpenedBackendRoutes' into OpenedRouteCriterion.
2) Add  'RouteBackendType' into RouteCriterionType.
3) Replace the Direction class with RouteType. The
direction is used to distinguish stream routes for
playback from those for capture. But now we need to
support for backend routes, and a backend route does
not has a direction since it has two ports with different
directions. So we need to use Routetype to distinguish 3
types of routes from each other: stream routes for
playback, stream routes for capture and backend routes.

Change-Id: Ia539dc256d76640af3b5707ad63c648e4a7c7de9
Tracked-On:
Signed-off-by: yuanwu <yuan.wu@intel.com>
Reviewed-by: Mengdong Lin <mengdong.lin@linux.intel.com>
Reviewed-by: Libin Yang <libin.yang@intel.com>