audio_hal: Apply the AudioBackendRoute class
Currently the audio backend routes are not parsed.
This patch creates an AudioBackendRoute instance
when a backend route element is parsed from the
audio_policy_configurations.xml. The RouteManagerConfig
saves these AudioBackendRoutes in the AudioRouteCollection,
so the AudioRouteManager can control both AudioBackendRoutes
and AudioStreamRoutes via the AudioRouteCollection.

Currently we have created the mapping relationship
between android parameters and audio backed routes.
But we need to enable/disable the audio backend routes
when the parameters is set by the setParameters function.

This patch set the route as selected, and enable/disable
the audio backend route.

Change-Id: Ic99d428c629e46a403257ff874914e151eb73c3b
Tracked-On:
Signed-off-by: yuanwu <yuan.wu@intel.com>
Reviewed-by: Mengdong Lin <mengdong.lin@linux.intel.com>
Reviewed-by: Libin Yang <libin.yang@intel.com>