audio_hal: Add the getRoute function into AudioRouteCollection
Currently we have created the mapping relationship
between android parameters and audio backed routes.
But the relationship only maps a parameter to the
name of an audio backend route and we still need to
find the AudioRoute pointer pointer by name from
the AudioRouteCollection, in order to control the
audio backend routes. This patch adds a getRoute
function to get the route by name.

Change-Id: I353051a08852f823f76cd533a1cbe8de4f451e0c
Tracked-On:
Signed-off-by: yuanwu <yuan.wu@intel.com>
Reviewed-by: Mengdong Lin <mengdong.lin@linux.intel.com>
Reviewed-by: Libin Yang <libin.yang@intel.com>